feat: adding meta tags to categories page and storybook enhancements
